pyproject.toml
[project]
name = "mcp-sysinfo"
version = "1.0.0"
description = "MCP server providing real-time system information"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "psutil>=5.9.0",
    "GPUtil>=1.4.0",
    "py-cpuinfo>=9.0.0",
]
